refactor : service 인터페이스화 및 Impl 로 분리
